feat: bitcoin funcionando
